% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{as.matrix,Deposits2-method}
\alias{as.matrix,Deposits2-method}
\alias{as.matrix,Deposits-method}
\alias{as.matrix,Deposits2D-method}
\alias{as.matrix,Trough-method}
\alias{as.matrix,Ellipsoid-method}
\alias{as.matrix,TrEllipsoid-method}
\alias{as.matrix,Sphere-method}
\alias{as.matrix,Cuboid-method}
\alias{as.matrix,Ellipse-method}
\alias{as.matrix,Ellipse-method}
\alias{as.matrix,TrEllipse-method}
\alias{as.matrix,Trough2D-method}
\alias{as.matrix,Rectangle-method}
\title{Conversion to matrix}
\usage{
\S4method{as.matrix}{Deposits2}(x)

\S4method{as.matrix}{Deposits}(x)

\S4method{as.matrix}{Deposits2D}(x)

\S4method{as.matrix}{Trough}(x)

\S4method{as.matrix}{Ellipsoid}(x)

\S4method{as.matrix}{TrEllipsoid}(x)

\S4method{as.matrix}{Sphere}(x)

\S4method{as.matrix}{Cuboid}(x)

\S4method{as.matrix}{Ellipse}(x)

\S4method{as.matrix}{Ellipse}(x)

\S4method{as.matrix}{TrEllipse}(x)

\S4method{as.matrix}{Trough2D}(x)

\S4method{as.matrix}{Rectangle}(x)
}
\description{
Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Conversion to matrix

Only vertical object > section from a TrEllipsoid/Trough

Only vertical object > section from a TrEllipsoid/Trough

Conversion to matrix
}
