% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRDM.R
\docType{package}
\name{CBRDM-package}
\alias{CBRDM}
\alias{CBRDM-package}
\title{CBRDM: Coarse, Braided River Deposit Model.}
\description{
The CBRDM package provides some S4 classes to represent the subsurface
heterogeneity of coarse, braided river deposits.
}
\section{3D classes}{

\itemize{
  \item \code{Trough}
  \item \code{Deposits}: consists of Trough and layers
}
}

\section{2D classes}{

\itemize{
  \item \code{Trough2D}
  \item \code{Deposits2D}
}
}
\references{
...
}

