% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{crossBedding}
\alias{crossBedding}
\title{crossBedding}
\usage{
crossBedding(x, para)
}
\arguments{
\item{x}{An object.}

\item{para}{A list with following elements:
\describe{
  \item{nF}{Number of cross-beds}
  \item{rpos}{Relative position from the trough center, between 0 and 1}
  \item{phi}{Angle of cross-beds in radian}
}}
}
\description{
Add a cross-bedding to an existing object
}
