% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{DepositsOld-class}
\alias{DepositsOld-class}
\title{An S4 class to represent coarse braided deposits.}
\description{
An instance of the class \code{DepositsOld} contains an instance of the class
\code{Trough} as well as the elevations of horizontal layers.
}
\section{Slots}{

\describe{
\item{\code{version}}{A character vector indicating the version of CBRDM}

\item{\code{troughs}}{An instance of the class \code{Trough}}

\item{\code{layers}}{A numeric vector corresponding to the layer elevations. This
independent of the throughs (e.g., if you take a section, it 
can happen that on some layers there are no objects).}

\item{\code{bbox}}{A length-three list defining the model boundary}
}}

\seealso{
\code{\link{Trough-class}}
}
