% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faciesProp}
\alias{faciesProp}
\title{Subsurface properties for each facies.}
\format{A list of three vectors. gp = poorly-sorted gravel,
        bm = bimodal gravel, ow = open-framework gravel
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
Huggenberger (1993), Jussel et al (1994), 
        Huber and Huggenberger (2016) doi:10.5194/hess-20-2035-2016
}
\usage{
faciesProp
}
\description{
Subsurface properties for each facies.
}
\keyword{datasets}

