% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{straussMH}
\alias{straussMH}
\title{Strauss process simulation (MCMC)}
\usage{
straussMH(bet = 10, gam = 0.5, d = 0.1, n0 = NULL, nit = 5000,
  W = list(x = c(0, 1), y = c(0, 1)), fd = NULL, count = FALSE)
}
\arguments{
\item{count}{boolean TRUE: return the number of points for each iteration}
}
\description{
strauss process: \eqn{bet^(n(y)) * gam^(s(y))}
with 0 <= gam <= 1 and bet > 0
if gam = 1, Strauss process = Poisson process
if gam = 0, Strauss process = Hard core process
}
