% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{straussMHGibbs}
\alias{straussMHGibbs}
\title{Strauss process (as Gibbs process) simulation (MCMC)}
\usage{
straussMHGibbs(alpha = 10, bet = 0.5, d = 0.1, n0 = NULL, nit = 5000,
  W = list(x = c(0, 1), y = c(0, 1)), fd = NULL, count = FALSE)
}
\arguments{
\item{count}{boolean TRUE: return the number of points for each iteration}
}
\description{
strauss process: exp(-(alpha*(n(y))  + beta*s(y)))
with alpha >= 0 and bet > 0
if beta = 0, Strauss process = Poisson process
if beta = +infinity, Strauss process = Hard core process
}
