% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{[[}
\alias{[[}
\alias{[[,Spoon-method}
\alias{[[,Trough2D-method}
\title{Subsetting}
\usage{
\S4method{[[}{Trough}(x, i, j, ...)

\S4method{[[}{Spoon}(x, i, j, ...)

\S4method{[[}{Trough2D}(x, i, j, ...)
}
\description{
Subsetting

Subsetting

Subsetting
}

