% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{as}
\alias{as}
\title{As("Ellipsoid", "TrEllipsoid")}
\description{
As("Ellipsoid", "TrEllipsoid")

As("TrEllipsoid", "Ellipsoid")

As("Trough", "TrEllipsoid")

As("Trough", "Ellipsoid")

As("TrEllipsoid", "Trough")

As("Ellipsoid", "Sphere")

As("Trough", "Sphere")

As("TrEllipsoid", "Sphere")

As("Trough", "Cuboid")

As("TrEllipsoid", "Cuboid")

As("Ellipsoid", "Cuboid")

As("TrEllipse", "Trough2D")

As("Trough2D", "TrEllipse")
}

