% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{Deposits-class}
\alias{Deposits-class}
\title{An S4 class to represent coarse braided deposits NEW.}
\description{
An instance of the class \code{Deposits} contains an instance of the class
\code{Trough} as well as the elevations of horizontal layers.
}
\section{Slots}{

\describe{
\item{\code{version}}{A character vector indicating the version of CBRDM}

\item{\code{layers}}{A list of instances of the class \code{Trough}}

\item{\code{z}}{A numeric vector corresponding to the layer elevations.}

\item{\code{bbox}}{A length-three list defining the model boundary}
}}

\seealso{
\code{\link{Trough-class}}
}
