% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{Ellipsoid-class}
\alias{Ellipsoid-class}
\title{An S4 class to represent Ellipsoid}
\description{
An S4 class to represent Ellipsoid
}
\section{Slots}{

\describe{
\item{\code{version}}{A length-n character vector indicating the version of RGPR}

\item{\code{id}}{A length-n numeric vector}

\item{\code{pos}}{A nx3 numeric matrix corresponding to object center position.}
}}

