% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{methods}
\name{[[}
\alias{[[}
\alias{[,Trough2D,ANY,ANY,ANY-method}
\title{Subsetting}
\usage{
\S4method{[}{Trough,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{Trough2D,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\description{
Subsetting

Subsetting
}
